fix #12

+++ src/main/java/ru/itpark/service/ProductService.java
@@ -88,8 +88,8 @@ public class ProductService {
         return result;
     }
 
-    public List<Product> pagingListItems(@NotNull List<Product> products, long countPerPage, int pageNumber) {
-        int countOfPages = (int)(products.size() / countPerPage);
+    public List<Product> pagingListItems(@NotNull List<Product> products, int countPerPage, int pageNumber) {
+        long countOfPages = products.size() / countPerPage;
         return products.stream().skip(Math.min(pageNumber - 1, countOfPages - 1) * countPerPage).limit(countPerPage).collect(Collectors.toList());
     }
 }
